requirements.txt
# Core dependencies
sphinx>=8.0.0
sphinx-rtd-theme>=3.0.0
python-dotenv>=1.1.0
lxml>=5.4.0
lxml_html_clean>=0.4.2

# Data processing and analysis
pandas>=2.0.0
numpy>=2.0.0
tqdm>=4.65.0

# NLP and text processing
langdetect>=1.0.9

# MLOps and model management
joblib>=1.3.0
mlflow>=2.8.0

# API interaction
fastapi>=0.115.12
uvicorn>=0.34.2

# Development tools
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

# Install project in editable mode
-e .
